requirements.txt
requests
psycopg2-binary
pytest-envfiles